pkgdown/custom-header.html

neuroSCC: Bridging Simultaneous Confidence Corridors with Applications and Performance Evaluation in PET Neuroimaging

Provides functions for the structured processing of PET neuroimaging data in preparation for the estimation of Simultaneous Confidence Corridors (SCCs) for one-group, two-group, or single patient vs group comparisons. The package facilitates PET image loading, data restructuring, integration into a Functional Data Analysis framework, contour extraction, post-SCC identification of significant results, and performance evaluation. It bridges established neuroimaging packages (e.g., 'oro.nifti') with novel statistical methodologies (i.e., 'ImageSCC') and facilitates reproducible analysis pipelines, enabling performance comparison of SCC-based analyses with gold-standard methods such as Statistical Parametric Mapping (SPM).

See also

Author

Maintainer: Juan A. Arias Lopez juanantonio.arias.lopez@usc.es

Other contributors:

Virgilio Gomez Rubio Virgilio.Gomez@uclm.es [reviewer]

Pablo Aguiar Fernandez pablo.aguiar@usc.es [thesis advisor, copyright holder]

Andrew Haddon Kemp A.H.Kemp@swansea.ac.uk [thesis advisor]

pkgdown/footer.html