Process ROIs from a NIfTI file

This function processes Regions of Interest (ROIs) from a binary NIfTI file. It extracts voxel coordinates from the image and preserves the original structure, marking which voxels are part of the ROI.

This function is typically used in SCC evaluation, where detected SCC regions are compared with predefined ROIs in calculateMetrics.

processROIs(
  roiFile,
  region,
  number,
  save = TRUE,
  outputDir = "results/ROIs",
  verbose = TRUE
)

Arguments

roiFile

character, the path to the NIfTI file containing the ROI data.

region

character, the name of the ROI region (e.g., "roi4").

number

character, the subject or group identifier (e.g., "18").

save

logical, if TRUE, saves the processed ROIs as `.RDS` files. If FALSE, prints a preview in the console. Default is TRUE.

outputDir

character, directory where processed ROI tables will be saved.

verbose

logical, if TRUE, prints progress messages. Default is TRUE.

Value

A data frame containing voxel-level ROI information, with columns:

group: ROI identifier, composed of region + number.

z, x, y: Voxel coordinates.

pet: Binary indicator (1 for ROI, 0 for non-ROI).

Details

The function reads the provided NIfTI file and extracts voxel data. It keeps all voxels, indicating whether each belongs to a ROI (pet = 1) or not (pet = 0).

See also

Examples

# Process an ROI NIfTI file (show results in console)
roiFile <- system.file("extdata", "ROIsample_Region2_18.nii", package = "neuroSCC")
processedROI <- processROIs(roiFile, region = "Region2", number = "18", save = FALSE)
#> Loading NIfTI file...
head(processedROI)  # Display first few rows
#>              group z x y pet
#> 1 Region2_number18 1 1 1   0
#> 2 Region2_number18 1 1 2   0
#> 3 Region2_number18 1 1 3   0
#> 4 Region2_number18 1 1 4   0
#> 5 Region2_number18 1 1 5   0
#> 6 Region2_number18 1 1 6   0